requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
moviepy==1.0.3
numpy==1.26.4
pillow==10.3.0
proglog==0.1.10
psycopg2==2.9.9
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
